version 2

+++ src/edu/berkeley/aep/Quantity.java
@@ -2,10 +2,10 @@ package edu.berkeley.aep;
 
 // Understands an amount in a given unit
 public class Quantity implements Bestable<Quantity> {
-    private final int size;
+    private final double size;
     private final Unit unit;
 
-    public Quantity(int size, Unit unit) {
+    public Quantity(double size, Unit unit) {
         this.size = size;
         this.unit = unit;
     }
@@ -25,7 +25,7 @@ public class Quantity implements Bestable<Quantity> {
         }
     }
 
-    private int convertTo(Unit toUnits) {
+    private double convertTo(Unit toUnits) {
         try {
             return unit.convertTo(toUnits, size);
         } catch (Exception e) {
@@ -40,7 +40,7 @@ public class Quantity implements Bestable<Quantity> {
 
     @Override
     public int hashCode() {
-        return Integer.hashCode(size);
+        return Double.hashCode(size);
     }
 
     public Quantity add(Quantity other) {
